Add files via upload
The objective of this project is to develop a stock price prediction model that uses previous stock prices along with news data to predict the stock prices of a set of companies. Mainly treating stock data as time-series, one can use past stock prices (and other parameters) to predict the stock prices for the next day or week.
To achieve this objective, we have collected a dataset of news articles and stock prices for a set of eighty-two companies. We then preprocess the news data to extract relevant features such as sentiment, topic, and source of news. Next, we will explore the correlation between news data and stock prices using statistical analysis and visualization techniques. We then use different machine learning algorithms to predict the future stock prices based on this consolidated dataset which comprise of both news data and stock price data.
In summary, this project aims to develop a stock price prediction model that includes a feature of news data along with the previous stock. Through this project, we hope to contribute to the development of new techniques for using news data along with trends in stock prices to predict future stock prices and improve the accuracy of stock price prediction models.